Send kills due to nick collisions back to the source.
This fixes a race condition with users changing hosts and then nick
colliding which causes servers to disagree on whether or not the
colliding users are the same user.

With this patch if this happens both users are killed instead, even if
their timestamps differ.

The proper way to fix this is to compare a 'real host', which we do not
store, or IP.

With TS5 and prior we could not do this because the KILLs would be
targeting nicks and not UIDs, which would not work as both sides would
always lose, but we can do this now as the other side can safely drop
the KILL for a nonexistant UID if the servers do agree.